refactor(overall): refactor code structure for improved readability and maintainability
