+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: "3"
-
 services:
   server:
     build:
@@ -7,13 +5,13 @@ services:
       dockerfile: Dockerfile
     restart: always
     container_name: server
-    command: uvicorn main:app --host 0.0.0.0 --port 5000 --reload
+    command: uvicorn app.main:app --host 0.0.0.0 --port 80
     healthcheck:
-      test: [ "CMD", "curl", "-f", "http://172.17.0.1:80" ]
+      test: ["CMD", "curl", "-f", "http://172.17.0.1:80"]
       interval: 1m
       timeout: 10s
       retries: 3
     ports:
-      - 80:5000
+      - 80:80
     volumes:
       - .:/app
