--- main.py
@@ -1,1 +1,0 @@
-from src.main import app
