+++ pyproject.toml
@@ -1,108 +1,111 @@
-[tool.poetry]
-name = "a2wsgi-examples"
+[project]
+name = "app"
 version = "0.1.0"
 description = ""
-authors = ["hasansezertasan <hasansezertasan@gmail.com>"]
 readme = "README.md"
-maintainers = ["hasansezertasan <hasansezertasan@gmail.com>"]
+authors = [
+    {name = "Hasan Sezer Taşan", email = "hasansezertasan@gmail.com"},
+]
+maintainers = [
+    {name = "Hasan Sezer Taşan", email = "hasansezertasan@gmail.com"},
+]
+requires-python = ">=3.10"
+dependencies = [
+    "a2wsgi>=1.10.10",
+    "blacksheep>=2.4.5",
+    "bottle>=0.13.4",
+    "django>=5.2.8",
+    "falcon>=4.2.0",
+    "fastapi>=0.121.3",
+    "flask>=3.1.2",
+    "flet[all]>=0.28.3",
+    "gradio>=5.50.0",
+    "litestar>=2.18.0",
+    "nicegui>=3.3.1",
+    "pyramid>=2.0.2",
+    "pywebio>=1.8.4",
+    "quart>=0.20.0",
+    "sanic>=25.3.0",
+    "uvicorn>=0.38.0",
+]
 
 
-[tool.poetry.dependencies]
-python = ">=3.11,<3.12"
-uvicorn = "^0.22.0"
-a2wsgi = "^1.7.0"
-FastAPI = "^0.101.0"
-Flask = "^3.0.0"
-Starlette = "^0.27.0"
-Django = "^4.2.6"
-bottle = "^0.12.25"
-litestar = "^2.1.1"
-gradio = "^3.47.1"
-pywebio = "^1.8.2"
-nicegui = "^1.3.16"
-flet-fastapi = "^0.10.3"
+[dependency-groups]
+dev = [
+  "ruff==0.14.6",
+  "ty==0.0.1a26",
+  "validate-pyproject[all]==0.24.1",
+  "vulture==2.14",
+  "typos==1.39.1",
+  "taplo==0.9.3",
+]
+test = [
+    "pytest==8.3.4",
+    "pytest-asyncio==0.25.3",
+]
 
 
-[tool.poetry.group.dev.dependencies]
-ipykernel = "^6.25.2"
-isort = "^5.12.0"
-black = { extras = ["jupyter"], version = "^23.9.1" }
-ruff = "^0.0.292"
-pre-commit = "^3.4.0"
-
-
-[tool.poetry.group.test.dependencies]
-pytest = "^7.4.0"
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 
 [tool.ruff]
-# Run `ruff linter` command to see all available linters and their descriptions
-target-version = "py311"
-line-length = 120
-select = [
-  # "B",
-  # "C",
-  "E",
-  "F",
-  "W",
-  "I",
-  "B9",
-]
-ignore = [
-  "B008", # do not perform function calls in argument defaults
-  "C901", # too complex
-  "E501", # line too long, handled by black
-  "F401", # imported but unused
-  "E402", # module level import not at top of file
-  "E712", # comparison to True should be 'if cond is True:' or 'if cond:'
-  "F821", # undefined name
-]
-exclude = ["alembic", "migrations"]
+fix = true
+output-format = "full"
+preview = true
 show-fixes = true
+src = ["src", "tests", "examples"]
+target-version = "py310"
+unsafe-fixes = true
 
 
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-use_parentheses = true
-ensure_newline_before_comments = true
-line_length = 120
-split_on_trailing_comma = true
-skip_glob = ["**/alembic/**", "**/migrations/**"]
+[tool.ruff.analyze]
+detect-string-imports = true
+preview = true
 
 
-[tool.black]
-line-length = 120
-target-version = ['py311']
-exclude = '''
-(
-  /(
-      \.eggs
-    | \.git
-    | \.hg
-    | \.mypy_cache
-    | \.nox
-    | \.tox
-    | \.venv
-    | _build
-    | buck-out
-    | build
-    | dist
-    | alembic
-  )/
-)
-'''
+[tool.ruff.format]
+docstring-code-format = true
+line-ending = "lf"
+preview = true
 
 
-[tool.poe.tasks]
-black = "black ."
-isort = "isort ."
-ruff = "ruff ."
-bir = ["black", "isort", "ruff"]
-line-count = "pygount --format=summary"
-test = "pytest -v -s"
-run = "uvicorn --port 5000 --reload main:app"
-clean = "frenchmaid clean"
-requirements = "poetry export -f requirements.txt --output pyproject.txt --without-hashes"
+[tool.ruff.lint]
+fixable = ["ALL"]
+ignore = ["COM812", "CPY", "FIX", "TD", "T201"]
+logger-objects = ["logging_setup.logger"]
+preview = true
+select = ["ALL"]
+
+
+[tool.ruff.lint.flake8-copyright]
+author = "Hasan Sezer Taşan <hasansezertasan@gmail.com>"
+
+
+[tool.ruff.lint.flake8-tidy-imports]
+ban-relative-imports = "all"
+
+
+[tool.ruff.lint.mccabe]
+max-complexity = 5
+
+
+[tool.ruff.lint.pycodestyle]
+ignore-overlong-task-comments = true
+
+
+[tool.ruff.lint.pydocstyle]
+convention = "pep257"
+
+
+[tool.ruff.lint.per-file-ignores]
+"tests/*.py" = ['S101']
+
+
+[tool.ruff.lint.pyupgrade]
+keep-runtime-typing = true
+
+
+[tool.ruff.lint.ruff]
+parenthesize-tuple-in-subscript = true
