+++ renovate.json
@@ -1,6 +1,7 @@
 {
   "$schema": "https://docs.renovatebot.com/renovate-schema.json",
-  "extends": [
-    "config:recommended"
-  ]
+  "dependencyDashboard": true,
+  "extends": ["config:recommended"],
+  "labels": ["internal"],
+  "semanticCommits": "enabled"
 }
