+++ src/app/applications/aiohttp/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of aiohttp application."""
