+++ src/app/applications/bottle/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of bottle application."""
