+++ src/app/applications/bottle/app.py
@@ -1,0 +1,10 @@
+from bottle import Bottle
+
+app = Bottle()
+
+
+@app.route("/")
+def hello():
+    return {
+        "message": "Hello bottle",
+    }
