+++ src/app/applications/django/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of django application."""
