+++ src/app/applications/fastapi/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of fastapi application."""
