+++ src/app/applications/flask/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of flask application."""
