+++ src/app/applications/flet/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of flet application."""
