+++ src/app/applications/gradio/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of gradio application."""
