+++ src/app/applications/litestar/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of litestar application."""
