+++ src/app/applications/litestar/app.py
@@ -1,0 +1,9 @@
+from litestar import Litestar, get
+
+
+@get("/")
+async def index() -> dict:
+    return {"message": "Hello, Litestar!"}
+
+
+app = Litestar([index])
