+++ src/app/applications/nicegui/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of nicegui application."""
