+++ src/app/applications/nicegui/frontend.py
