+++ src/app/applications/pywebio/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of pywebio application."""
