+++ src/app/applications/starlette/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation of starlette application."""
