--- src/applications/bottle/app.py
@@ -1,10 +1,0 @@
-from bottle import Bottle
-
-app = Bottle()
-
-
-@app.route("/")
-def hello():
-    return {
-        "message": "Hello World",
-    }
