--- src/applications/litestar/app.py
@@ -1,9 +1,0 @@
-from litestar import Litestar, get
-
-
-@get("/")
-async def index() -> dict:
-    return {"message": "Hello, World!"}
-
-
-app = Litestar([index])
