--- src/applications/starlette/app.py
@@ -1,19 +1,0 @@
-from starlette.applications import Starlette
-from starlette.responses import JSONResponse
-from starlette.routing import Route
-
-
-async def homepage(request):
-    return JSONResponse(
-        {
-            "message": "Hello World",
-        },
-    )
-
-
-app = Starlette(
-    debug=True,
-    routes=[
-        Route("/", homepage),
-    ],
-)
